Move SharedCode utilities to WidgetAPIHttpTrigger function
Only one function in function app
