+++ WidgetAPIHttpTrigger/tests/test_valid_params.py
@@ -1,13 +1,6 @@
 import unittest
-import os
-import sys
-import inspect
 
 from course_param_validator import valid_course_params
-
-CURRENTDIR = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-PARENTDIR = os.path.dirname(CURRENTDIR)
-sys.path.insert(0, PARENTDIR)
 
 
 class TestValidCourseParams(unittest.TestCase):
